boxshop/requirements.txt
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
boto3==1.17.54
brotlipy==0.7.0
certifi==2021.10.8
clyent==1.2.1
django-cors-headers==3.13.0
django-default-imagefield==0.5.5
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
fonttools==4.25.0
gunicorn==20.1.0
inflection==0.5.1
llvmlite==0.38.0
mccabe==0.6.1
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
pandas==1.4.2
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==9.0.1
protobuf==3.19.1
psycopg2==2.9.3
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.43.0.6
PyDispatcher==2.0.5
pyls-spyder==0.4.0
pytest==7.1.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
pytz==2021.3
sqlparse==0.4.2
tabulate==0.8.9
whitenoise==5.1.0